requirements.txt
aioamqp==0.15.0
aiohttp==3.9.3
aioresponses==0.7.8
alembic==1.13.1
annotated-types==0.7.0
asyncpg==0.29.0
bitarray==3.4.3
bitcoinlib==0.7.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
cryptography==45.0.5
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastecdsa==3.0.1
flake8==6.1.0
freeze-uuid==0.4.3
freezegun==1.2.2
greenlet==3.2.3
hexbytes==1.3.1
httpx==0.27.0
import-linter==2.1
iniconfig==2.0.0
isort==5.12.0
lxml==4.9.4
mypy==1.5.1
mypy-extensions==1.0.0
numpy==1.26.4
parsimonious==0.10.0
Pillow==11.3.0
pre-commit==3.3.3
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pytest==7.4.0
pytest-aiohttp==1.0.4
pytest-alembic==0.10.7
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-env==1.0.0
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pyunormalize==16.0.0
qrcode==8.2
regex==2024.11.6
requests==2.32.4
respx==0.21.1
rlp==4.1.0
SQLAlchemy==2.0.28
toolz==1.0.0
types-python-dateutil==2.8.19.14
types-requests==2.32.4.20250611
types-ujson==5.9.0.0
typing-inspection==0.4.1
ujson==5.9.0
urllib3==2.5.0
uvloop==0.17.0
web3==7.12.0
websockets==15.0.1
